fix: changing attribute type and removing optional

+++ src/main/java/com/jaya/challenge/api/currency/converter/repository/UserRepository.java
@@ -13,10 +13,10 @@ import com.jaya.challenge.api.currency.converter.model.entity.User;
  *
  */
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 	
 	@Query("select u from User u where u.idUser = ?1")
-	Optional<User> findById(Integer idUser);
+	Optional<User> findById(Long idUser);
 
 
 }
